rename Miners to Mining folder and namespace
