+++ src/NiceHashMiner/Algorithms/PluginAlgorithm.cs
@@ -1,9 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using NiceHashMiner.Miners.IntegratedPlugins;
 using NiceHashMiner.Plugin;
 using NHM.Common.Enums;
 using AlgorithmCommon = NHM.Common.Algorithm;
