+++ src/NiceHashMiner/ApplicationStateManager/ApplicationStateManager.DataVizualizers.cs
@@ -1,11 +1,5 @@
-using NiceHashMiner.Configs;
 using NiceHashMiner.Interfaces.DataVisualizer;
-using NiceHashMiner.Stats;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace NiceHashMiner
 {
