+++ src/NiceHashMiner/ApplicationStateManager/ApplicationStateManager.Devices.cs
@@ -1,11 +1,7 @@
 using NiceHashMiner.Devices;
 using NiceHashMiner.Stats;
-using NHM.Common.Enums;
-using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace NiceHashMiner
 {
