+++ src/NiceHashMiner/ApplicationStateManager/ApplicationStateManager.Mining.cs
@@ -6,7 +6,7 @@ using System.Text;
 using System.Threading.Tasks;
 using NHM.Common.Enums;
 using NiceHashMiner.Benchmarking;
-using NiceHashMiner.Miners;
+using NiceHashMiner.Mining;
 using NiceHashMiner.Stats;
 using NHM.Common;
 using NiceHashMiner.Configs;
