+++ src/NiceHashMiner/ApplicationStateManager/ApplicationStateManager.Timers.cs
@@ -1,7 +1,7 @@
 using NHM.DeviceDetection;
 using NiceHashMiner.Configs;
 using NiceHashMiner.Devices;
-using NiceHashMiner.Miners;
+using NiceHashMiner.Mining;
 using NHM.Common;
 using System;
 using System.Diagnostics;
