+++ src/NiceHashMiner/ApplicationStateManager/ApplicationStateManager.cs
@@ -1,17 +1,11 @@
 using NHM.UUID;
 using NiceHashMiner.Configs;
 using NiceHashMiner.Devices;
-using NiceHashMiner.Interfaces.DataVisualizer;
-using NiceHashMiner.Miners;
+using NiceHashMiner.Mining;
 using NiceHashMiner.Stats;
-using NHM.Common;
 using NHM.Common.Enums;
 using System;
-using System.Collections.Generic;
-using System.Diagnostics;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
 using NiceHashMiner.Utils;
