+++ src/NiceHashMiner/Benchmarking/BenchmarkManager.cs
@@ -2,13 +2,13 @@ using NiceHashMiner.Algorithms;
 using NHM.Extensions;
 using NiceHashMiner.Devices;
 using NiceHashMiner.Interfaces;
-using NiceHashMiner.Miners;
+using NiceHashMiner.Mining;
 using NHM.Common.Enums;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using NiceHashMiner.Stats;
-using NiceHashMiner.Miners.IntegratedPlugins;
+using NiceHashMiner.Mining.IntegratedPlugins;
 
 namespace NiceHashMiner.Benchmarking
 {
