+++ src/NiceHashMiner/Configs/Data/GeneralConfig.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Drawing;
-using NiceHashMiner.Miners;
+using NiceHashMiner.Mining;
 using NiceHashMiner.Switching;
 using NHM.Common.Enums;
 using NiceHashMiner.Utils;
