+++ src/NiceHashMiner/Devices/AvailableDevices.cs
@@ -1,5 +1,4 @@
-﻿using NHM.Common;
-using NHM.Common.Device;
+﻿using NHM.Common.Device;
 using NHM.Common.Enums;
 using System.Collections.Generic;
 using System.Linq;
