+++ src/NiceHashMiner/Forms/Form_Benchmark.cs
@@ -9,8 +9,8 @@ using NiceHashMiner.Benchmarking;
 using NiceHashMiner.Configs;
 using NiceHashMiner.Devices;
 using NiceHashMiner.Interfaces;
-using NiceHashMiner.Miners;
-using NiceHashMiner.Miners.Grouping;
+using NiceHashMiner.Mining;
+using NiceHashMiner.Mining.Grouping;
 using NiceHashMiner.Properties;
 using NHM.Common.Enums;
 using Timer = System.Windows.Forms.Timer;
