+++ src/NiceHashMiner/Mining/IntegratedPlugins/CCMinerIntegrated/CCMinerTpruvotIntegratedPlugin.cs
@@ -9,7 +9,7 @@ using NHM.Common.Algorithm;
 using NHM.Common.Device;
 using NHM.Common.Enums;
 
-namespace NiceHashMiner.Miners.IntegratedPlugins
+namespace NiceHashMiner.Mining.IntegratedPlugins
 {
     class CCMinerTpruvotIntegratedPlugin : CCMinersPluginBase, IReBenchmarkChecker
     {
