+++ src/NiceHashMiner/Mining/IntegratedPlugins/IntegratedPluginClasses.cs
@@ -1,5 +1,5 @@
 ﻿
-namespace NiceHashMiner.Miners.IntegratedPlugins
+namespace NiceHashMiner.Mining.IntegratedPlugins
 {
     class ClaymoreDual14IntegratedPlugin : ClaymoreDual14.ClaymoreDual14Plugin, IntegratedPlugin
     {
