+++ src/NiceHashMiner/Mining/IntegratedPlugins/MinersBinsUrls.cs
@@ -4,7 +4,7 @@ using NHM.Common;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace NiceHashMiner.Miners.IntegratedPlugins
+namespace NiceHashMiner.Mining.IntegratedPlugins
 {
     internal static class MinersBinsUrls
     {
