+++ src/NiceHashMiner/Mining/IntegratedPlugins/SGminerIntegratedHub/SGminerIntegratedMiner.cs
@@ -10,7 +10,7 @@ using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace NiceHashMiner.Miners.IntegratedPlugins
+namespace NiceHashMiner.Mining.IntegratedPlugins
 {
     class SGminerIntegratedMiner : SGMinerBase
     {
