+++ src/NiceHashMiner/Mining/Miner.cs
