+++ src/NiceHashMiner/NiceHashMiner.csproj
@@ -237,32 +237,32 @@
     <Compile Include="Interfaces\IRatesComunication.cs" />
     <Compile Include="Interfaces\StateSetters\IEnabledDeviceStateSetter.cs" />
     <Compile Include="Interfaces\StateSetters\IStateSetter.cs" />
-    <Compile Include="Miners\IdleChecking\IdleChangedEventArgs.cs" />
-    <Compile Include="Miners\IdleChecking\IdleChecker.cs" />
-    <Compile Include="Miners\IdleChecking\IdleCheckManager.cs" />
-    <Compile Include="Miners\IdleChecking\InputTimeoutChecker.cs" />
-    <Compile Include="Miners\IdleChecking\SessionLockChecker.cs" />
-    <Compile Include="Miners\IntegratedPlugins\CCMinerIntegrated\CCMinerIntegratedMiner.cs" />
-    <Compile Include="Miners\IntegratedPlugins\CCMinerIntegrated\CCMinerMTPIntegratedPlugin.cs" />
-    <Compile Include="Miners\IntegratedPlugins\CCMinerIntegrated\CCMinersPluginBase.cs" />
-    <Compile Include="Miners\IntegratedPlugins\CCMinerIntegrated\CCMinerTpruvotIntegratedPlugin.cs" />
-    <Compile Include="Miners\IntegratedPlugins\IntegratedPlugin.cs" />
-    <Compile Include="Miners\IntegratedPlugins\IntegratedPluginClasses.cs" />
-    <Compile Include="Miners\IntegratedPlugins\MinersBinsUrls.cs" />
-    <Compile Include="Miners\IntegratedPlugins\SGminerIntegratedHub\SGminerAvemoreIntegratedPlugin.cs" />
-    <Compile Include="Miners\IntegratedPlugins\SGminerIntegratedHub\SGminerGMIntegratedPlugin.cs" />
-    <Compile Include="Miners\IntegratedPlugins\SGminerIntegratedHub\SGminerIntegratedMiner.cs" />
-    <Compile Include="Miners\IntegratedPlugins\SGminerIntegratedHub\SGMinerPluginBase.cs" />
-    <Compile Include="Miners\IntegratedPlugins\VC_REDIST_x64_2015_DEPENDENCY_PLUGIN.cs" />
-    <Compile Include="Miners\MinerApiWatchdog.cs" />
-    <Compile Include="Miners\MiningManager.cs" />
-    <Compile Include="Miners\MiningStats.cs" />
+    <Compile Include="Mining\IdleChecking\IdleChangedEventArgs.cs" />
+    <Compile Include="Mining\IdleChecking\IdleChecker.cs" />
+    <Compile Include="Mining\IdleChecking\IdleCheckManager.cs" />
+    <Compile Include="Mining\IdleChecking\InputTimeoutChecker.cs" />
+    <Compile Include="Mining\IdleChecking\SessionLockChecker.cs" />
+    <Compile Include="Mining\IntegratedPlugins\CCMinerIntegrated\CCMinerIntegratedMiner.cs" />
+    <Compile Include="Mining\IntegratedPlugins\CCMinerIntegrated\CCMinerMTPIntegratedPlugin.cs" />
+    <Compile Include="Mining\IntegratedPlugins\CCMinerIntegrated\CCMinersPluginBase.cs" />
+    <Compile Include="Mining\IntegratedPlugins\CCMinerIntegrated\CCMinerTpruvotIntegratedPlugin.cs" />
+    <Compile Include="Mining\IntegratedPlugins\IntegratedPlugin.cs" />
+    <Compile Include="Mining\IntegratedPlugins\IntegratedPluginClasses.cs" />
+    <Compile Include="Mining\IntegratedPlugins\MinersBinsUrls.cs" />
+    <Compile Include="Mining\IntegratedPlugins\SGminerIntegratedHub\SGminerAvemoreIntegratedPlugin.cs" />
+    <Compile Include="Mining\IntegratedPlugins\SGminerIntegratedHub\SGminerGMIntegratedPlugin.cs" />
+    <Compile Include="Mining\IntegratedPlugins\SGminerIntegratedHub\SGminerIntegratedMiner.cs" />
+    <Compile Include="Mining\IntegratedPlugins\SGminerIntegratedHub\SGMinerPluginBase.cs" />
+    <Compile Include="Mining\IntegratedPlugins\VC_REDIST_x64_2015_DEPENDENCY_PLUGIN.cs" />
+    <Compile Include="Mining\MinerApiWatchdog.cs" />
+    <Compile Include="Mining\MiningManager.cs" />
+    <Compile Include="Mining\MiningStats.cs" />
     <Compile Include="NHMProductInfo.cs" />
     <Compile Include="Plugin\PluginContainer.cs" />
     <Compile Include="Plugin\SupportedAlgorithmsFilter.cs" />
     <Compile Include="Plugin\SupportedPluginsFilter.cs" />
     <Compile Include="Plugin\MinerPluginsManager.cs" />
-    <Compile Include="Miners\MinerFromPlugin.cs" />
+    <Compile Include="Mining\MinerFromPlugin.cs" />
     <Compile Include="Stats\ErrorCode.cs" />
     <Compile Include="Stats\Models\ExchangeRateJson.cs" />
     <Compile Include="Stats\Models\MarketsMessage.cs" />
@@ -339,13 +339,13 @@
     <Compile Include="Interfaces\IBenchmarkCalculation.cs" />
     <Compile Include="Interfaces\IListItemCheckColorSetter.cs" />
     <Compile Include="Utils\Links.cs" />
-    <Compile Include="Miners\Grouping\GroupSetupUtils.cs" />
-    <Compile Include="Miners\Grouping\MiningDevice.cs" />
-    <Compile Include="Miners\MinerFactory.cs" />
-    <Compile Include="Miners\Grouping\GroupingUtils.cs" />
+    <Compile Include="Mining\Grouping\GroupSetupUtils.cs" />
+    <Compile Include="Mining\Grouping\MiningDevice.cs" />
+    <Compile Include="Mining\MinerFactory.cs" />
+    <Compile Include="Mining\Grouping\GroupingUtils.cs" />
     <Compile Include="Utils\CredentialValidators.cs" />
     <Compile Include="Utils\Helpers.cs" />
-    <Compile Include="Miners\Miner.cs" />
+    <Compile Include="Mining\Miner.cs" />
     <Compile Include="Forms\MinerFileNotFoundDialog.cs">
       <SubType>Form</SubType>
     </Compile>
