+++ src/NiceHashMiner/Stats/NiceHashStats_production_new.cs
@@ -14,7 +14,7 @@ using WebSocketSharp;
 // static imports
 using static NiceHashMiner.Stats.StatusCodes;
 using NiceHashMiner.Utils;
-using NiceHashMiner.Miners;
+using NiceHashMiner.Mining;
 
 namespace NiceHashMiner.Stats
 {
