+++ src/NiceHashMiner/Stats/NiceHashStats_production_old.cs
@@ -3,7 +3,7 @@
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using NiceHashMiner.Devices;
-using NiceHashMiner.Miners;
+using NiceHashMiner.Mining;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
